Parser memory-safe rewrite wip
