+++ old/Codegen/codegen.cc
@@ -2,7 +2,6 @@
 #include "../../lib/error.h"
 #include "../../lib/list.h"
 #include "../../lib/logger.h"
-#include "../AST_old/ast.hh"
 #include "llvm/IR/IRBuilder.h"
 #include "llvm/IR/Module.h"
 #include "llvm/IR/Type.h"
