+++ old/Parser_old/parser.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "../token/token.h"
+#include <vector>
+namespace Parser
+{
+CompilationUnit* makeAst(std::vector<Token>& tokens);
+}
