+++ old/Processor/passes/include.h
@@ -1,6 +1,5 @@
 #pragma once
 #include<string>
 #include<vector>
-#include "../../AST_old/ast.hh"
 #include "../../../lib/logger.h"
 #include "../../../lib/list.h"
