+++ old/Processor/passes/required/compositor_resolver.hh
