+++ old/Processor/passes/required/property_resolver.hh
