+++ old/Processor/passes/sugar/self_ref_resolver.hh
