+++ old/Processor/processor.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "../AST_old/ast.hh"
 namespace Processor
 {
 void process(CompilationUnit* unit, bool finalize = true);
